Update the createGrid func so its not hardcoded.
The change allows the the height and width of the div to perfectly fit -
inside of the container by dividing the width of the container to the -
grid size. The grid also clears itself using innerHTML to stop new -
squares adding to the existing ones when calling the func each time.